+++ tests/trainer/test_lr_finder.py
@@ -78,8 +78,8 @@ def test_trainer_reset_correctly(tmpdir):
 def test_trainer_arg_bool(tmpdir):
     """ Test that setting trainer arg to bool works """
     hparams = EvalModelTemplate.get_default_hparams()
-    model = EvalModelTemplate(hparams)
-    before_lr = hparams.learning_rate
+    model = EvalModelTemplate(**hparams)
+    before_lr = hparams.get('learning_rate')
 
     # logger file to get meta
     trainer = Trainer(
@@ -89,18 +89,17 @@ def test_trainer_arg_bool(tmpdir):
     )
 
     trainer.fit(model)
-    after_lr = model.hparams.learning_rate
+    after_lr = model.learning_rate
     assert before_lr != after_lr, \
         'Learning rate was not altered after running learning rate finder'
 
 
 def test_trainer_arg_str(tmpdir):
     """ Test that setting trainer arg to string works """
-    hparams = EvalModelTemplate.get_default_hparams()
-    hparams.__dict__['my_fancy_lr'] = 1.0  # update with non-standard field
-    model = EvalModelTemplate(hparams)
+    model = EvalModelTemplate()
+    model.my_fancy_lr = 1.0  # update with non-standard field
 
-    before_lr = hparams.my_fancy_lr
+    before_lr = model.my_fancy_lr
     # logger file to get meta
     trainer = Trainer(
         default_save_path=tmpdir,
@@ -109,7 +108,7 @@ def test_trainer_arg_str(tmpdir):
     )
 
     trainer.fit(model)
-    after_lr = model.hparams.my_fancy_lr
+    after_lr = model.my_fancy_lr
     assert before_lr != after_lr, \
         'Learning rate was not altered after running learning rate finder'
 
@@ -118,18 +117,18 @@ def test_call_to_trainer_method(tmpdir):
     """ Test that directly calling the trainer method works """
 
     hparams = EvalModelTemplate.get_default_hparams()
-    model = EvalModelTemplate(hparams)
+    model = EvalModelTemplate(**hparams)
 
-    before_lr = hparams.learning_rate
+    before_lr = hparams.get('learning_rate')
     # logger file to get meta
     trainer = Trainer(
         default_save_path=tmpdir,
-        max_epochs=5
+        max_epochs=5,
     )
 
     lrfinder = trainer.lr_find(model, mode='linear')
     after_lr = lrfinder.suggestion()
-    model.hparams.learning_rate = after_lr
+    model.learning_rate = after_lr
     trainer.fit(model)
 
     assert before_lr != after_lr, \
@@ -141,9 +140,9 @@ def test_accumulation_and_early_stopping(tmpdir):
         accumulation also works for this feature """
 
     hparams = EvalModelTemplate.get_default_hparams()
-    model = EvalModelTemplate(hparams)
+    model = EvalModelTemplate(**hparams)
 
-    before_lr = hparams.learning_rate
+    before_lr = hparams.get('learning_rate')
     # logger file to get meta
     trainer = Trainer(
         default_save_path=tmpdir,
@@ -165,12 +164,12 @@ def test_suggestion_parameters_work(tmpdir):
     """ Test that default skipping does not alter results in basic case """
 
     hparams = EvalModelTemplate.get_default_hparams()
-    model = EvalModelTemplate(hparams)
+    model = EvalModelTemplate(**hparams)
 
     # logger file to get meta
     trainer = Trainer(
         default_save_path=tmpdir,
-        max_epochs=10
+        max_epochs=10,
     )
 
     lrfinder = trainer.lr_find(model)
@@ -200,3 +199,26 @@ def test_suggestion_with_non_finite_values(tmpdir):
 
     assert before_lr == after_lr, \
         'Learning rate was altered because of non-finite loss values'
+
+
+def test_logger_reset_correctly(tmpdir):
+    """ Test that logger is updated correctly """
+    tutils.reset_seed()
+
+    hparams = EvalModelTemplate.get_default_hparams()
+    model = EvalModelTemplate(hparams)
+
+    trainer = Trainer(
+        default_save_path=tmpdir,
+        max_epochs=10,
+        auto_lr_find=True
+    )
+    logger1 = trainer.logger
+    trainer.fit(model)
+    logger2 = trainer.logger
+    logger3 = model.logger
+
+    assert logger1 == logger2, \
+        'Learning rate finder altered the logger of trainer'
+    assert logger2 == logger3, \
+        'Learning rate finder altered the logger of model'